pyproject.toml
[tool.poetry]
name = "whisper-api"
version = "0.1.0"
description = "Backend for a transcribe-to-print pipeline called SessionScribe"
authors = ["Christopher Carroll Smith <75859865+chriscarrollsmith@users.noreply.github.com>"]
readme = "README.md"
packages = [
  {include = "api"},
  {include = "tests"}
]

[tool.poetry.dependencies]
python = "^3.9"
dacite = "^1.8.1"
jiwer = "^3.0.2"
ffmpeg-python = "^0.2.0"
gql = {extras = ["all"], version = "^3.4.1"}
pandas = "^2.0.3"
loguru = "^0.7.0"
torchaudio = "^2.0.2"
yt-dlp = "^2023.7.6"
fpdf = "^1.7.2"
google-cloud-storage = "^2.10.0"
modal = "^0.50.2668"
supabase = "^1.0.3"
pytz = "^2023.3"
whisper = "^1.1.10"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
python-dotenv = "^1.0.0"
epson-connect = "^0.2.3"
uvicorn = "^0.23.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
